requirements.txt
# ========================================
# GARCH ALGO INTELLIGENCE PLATFORM
# Requirements - Pinned Versions
# ========================================

# Core Web Framework
streamlit==1.29.0
streamlit-autorefresh==1.0.1

# Data Manipulation & Analysis
pandas==2.1.4
numpy==1.26.2

# Financial Data Providers
yfinance==0.2.33
ccxt==4.2.3
requests==2.31.0

# Time Series & Econometric Models
arch==6.3.0
statsmodels==0.14.1

# Machine Learning & AI
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
prophet==1.1.5
xgboost==2.0.3

# Visualization
matplotlib==3.8.2
mplfinance==0.12.10b0
plotly==5.18.0
seaborn==0.13.0

# Technical Analysis
ta==0.11.0
pandas-ta==0.3.14b0

# PDF Generation & Reporting
reportlab==4.0.8
fpdf2==2.7.7
matplotlib-venn==0.11.9

# Data Validation
pydantic==2.5.3
cerberus==1.3.5

# Environment & Configuration
python-dotenv==1.0.0

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.39.2

# Database (Optional - for alerts/trades storage)
sqlalchemy==2.0.25
alembic==1.13.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
tqdm==4.66.1
joblib==1.3.2

# Performance
numba==0.58.1
bottleneck==1.3.7

# Alerts & Notifications
smtplib2==0.1.0
twilio==8.11.1
discord-webhook==1.3.1

# HTTP & API
httpx==0.26.0
aiohttp==3.9.1

# Type Hints
typing-extensions==4.9.0
